+++ libraries/INA228/CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.2.0] - 2025-03-05
+- fix #20, INA.getCharge => make it signed.
+- add **getLastError()** low level (I2C) error handling
+
+----
+
 ## [0.1.5] - 2025-01-30
 - fix #17, cache ADCRange to improve getShuntVoltage()
 - add **INA_comparison_table.md**